requirements.txt
torch
torchvision
requests
Pillow
numpy